test: fold $dist tests into the shared paginate.spec schema
distance.spec.ts stood up a second real-DB DataSource with its own dropSchema on the
same `test` database. Under parallel jest workers it raced with paginate.spec's
dropSchema — each suite dropped the other's tables mid-run (`Table 'test.cat'/'test.place'
doesn't exist`), only hidden locally by --runInBand.

Register PlaceEntity in paginate.spec's single schema and move the $dist tests into a
nested describe that seeds places off the shared connection, so there is one schema
loaded once at setup and no competing dropSchema. Parallel-safe on postgres, mariadb and
sqlite.
